Fixing actions (#1)
* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* fix narrowing conversion in text render command

* add -fpic option, as well as add test command in cmake file

* change c flag, and fix ctest

* Update cmake-multi-platform.yml

* change from clang to gcc

* fixed minor gcc bugs

* fix some includes

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* fix tests (hopefully)

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

* Update cmake-multi-platform.yml

+++ Source/RenderingGL/TextRenderCommand.cpp
@@ -61,7 +61,7 @@ TextRenderCommand::TextRenderCommand(uint32_t globalID, std::string text){
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x
+                static_cast<uint32_t>(face->glyph->advance.x)
             };
             g_characters.insert(std::pair<char, sCharacter>(c, character));
         }
